Second Pass
We had to fix some problems with the database.  Oh, and also add an actual main.ts file.  Afterward we tried to get google auth to work, but ran into some issues.  To save time poking at something I don't care much about, we switched to a simple auth scheme.

We have a lot of noise in the filesystem, but I suspect that we're in good shape to start the flashcards.  There will be an eventual reckoning, though :)