We've added verb conjugation
This was a tricky one.  Claude wanted to create an engine for the various verb families, but so many verbs are irregularly irregular, which implied a hybrid approach that was too brittle in practice.  So we have a more manual approach.

I might ask it to add more verbs manually later, but this will work for now.

I've decided to just include all file that might be garbage, and clean up the project at the end.